Extract Body files inline.
